backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
redis
httpx
cryptography
python-dotenv
email-validator
numpy
pandas
scikit-learn
joblib
xgboost
torch
psycopg2-binary
